+++ app/entities/City.js
@@ -1,4 +1,7 @@
 const { EntitySchema } = require('typeorm');
+const {
+  db: { srid },
+} = require('config');
 
 module.exports = new EntitySchema({
   name: 'City',
@@ -12,11 +15,12 @@ module.exports = new EntitySchema({
     name: {
       type: String,
     },
-    // geom: {
-    //   type: 'geometry',
-    //   spatialFeatureType: 'Point',
-    //   srid: +process.env.SRID,
-    //   nullable: false,
-    // },
+    geom: {
+      type: 'geometry',
+      spatialFeatureType: 'MultiPolygon',
+      srid,
+      nullable: false,
+      select: false,
+    },
   },
 });